// src/utils/enums.ts
export enum RouteEnum {
  Dashboard = '/',
  Theme = '/theme',
  Todo = '/todo', 
  Student = '/student',
}

export enum ThemeEnum {
  Light = 'light', 
  Dark = 'dark',
  Gray = 'gray',
}

export enum ActionEnum {
  ADD = 'ADD',
  TOGGLE = 'TOGGLE',
  DELETE = 'DELETE',
  EDIT = 'EDIT',
  INIT = 'INIT',
} 

export enum GenderEnum {  
 Male = "Male",
 Female = "Female",
 Other = "Other", 
}

export enum ClassEnum {
  A = "A",
  B = "B",
  C = "C",
}

export enum LocalStorageEnum {
  Todos = "todos",
  Students = "students",
  Theme = "app-theme", 
}